refactor vector tap into library
