+++ lib/vector-tap/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "vector-tap"
+version = "0.1.0"
+authors = ["Vector Contributors <vector@datadoghq.com>"]
+edition = "2021"
+publish = false
+license = "MPL-2.0"
+
+[dependencies]
+colored = { version = "2.1.0", default-features = false }
+serde_yaml = { version = "0.9.34", default-features = false }
+snafu = { version = "0.7.5", default-features = false }
+tokio = { version = "1.38.0", default-features = false, features = ["time"] }
+tokio-stream = { version = "0.1.15", default-features = false, features = ["sync"] }
+tracing = { version = "0.1.34", default-features = false }
+url = { version = "2.5.2", default-features = false }
+vector-api-client = {path = "../vector-api-client" }
